src/requirements.txt
mysqlclient==2.1.1
Django==4.0.5
djangorestframework==3.13.1
Markdown==3.3.7
django-filter==22.1
tzdata==2022.1
requests==2.28.0